--- .prettierrc.js
@@ -1,7 +1,0 @@
-module.exports =  {
-  semi:  true,
-  trailingComma:  'all',
-  singleQuote:  false,
-  printWidth:  120,
-  tabWidth:  4,
-};
