+++ index.js
@@ -1,20 +1,8 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-var btcClient_1 = require("./clients/btcClient");
-exports.btcClient = btcClient_1.client;
-var lncClient_1 = require("./clients/lncClient");
-exports.lnClient = lncClient_1.client;
-var otsClient_1 = require("./clients/otsClient");
-exports.otsClient = otsClient_1.client;
-var cypherNodeMatrixBridge_1 = require("./bridge/cypherNodeMatrixBridge");
-exports.cypherNodeMatrixBridge = cypherNodeMatrixBridge_1.cypherNodeMatrixBridge;
-var mqttUtil_1 = require("./lib/mqttUtil");
-exports.cypherNodeMqttSub = mqttUtil_1.cypherNodeMqttSub;
-var matrixUtil_1 = require("./lib/matrixUtil");
-exports.getSyncMatrixClient = matrixUtil_1.getSyncMatrixClient;
-var cypherNodeClient_1 = require("./lib/cypherNodeClient");
-exports.cypherNodeClient = cypherNodeClient_1.default;
-var cryptoUtil_1 = require("./lib/cryptoUtil");
-exports.cryptoUtils = cryptoUtil_1.crypto;
-var cyphernodeMatrixTransport_1 = require("./transport/cyphernodeMatrixTransport");
-exports.cypherNodeMatrixTransport = cyphernodeMatrixTransport_1.cypherNodeMatrixTransport;
+export { client as btcClient } from "./src/clients/btcClient";
+export { client as lnClient } from "./src/clients/lncClient";
+export { client as otsClient } from "./src/clients/otsClient";
+export { cypherNodeMatrixBridge } from "./src/bridge/cypherNodeMatrixBridge";
+export { getSyncMatrixClient } from "./src/lib/matrixUtil";
+export { default as cypherNodeClient } from "./src/transport/cypherNodeHttpTransport";
+export { crypto as cryptoUtils } from "./src/lib/cryptoUtil";
+export { cypherNodeMatrixTransport } from "./src/transport/cyphernodeMatrixTransport";
