+++ index.ts
@@ -1,11 +1,12 @@
-export { client as btcClient } from "./clients/btcClient";
-export { client as lnClient } from "./clients/lncClient";
-export { client as otsClient } from "./clients/otsClient";
-export { cypherNodeMatrixBridge } from "./bridge/cypherNodeMatrixBridge";
-export { cypherNodeMqttSub } from "./lib/mqttUtil";
-export { getSyncMatrixClient } from "./lib/matrixUtil";
-export { default as cypherNodeClient } from "./lib/cypherNodeClient";
-export { crypto as cryptoUtils } from "./lib/cryptoUtil";
+export { client as btcClient } from "./src/clients/btcClient";
+export { client as lnClient } from "./src/clients/lncClient";
+export { client as otsClient } from "./src/clients/otsClient";
+export { cypherNodeMatrixBridge } from "./src/bridge/cypherNodeMatrixBridge";
+export { getSyncMatrixClient } from "./src/lib/matrixUtil";
+export {
+  default as cypherNodeClient
+} from "./src/transport/cypherNodeHttpTransport";
+export { crypto as cryptoUtils } from "./src/lib/cryptoUtil";
 export {
   cypherNodeMatrixTransport
-} from "./transport/cyphernodeMatrixTransport";
+} from "./src/transport/cyphernodeMatrixTransport";
