+++ src/bridge/cypherNodeMatrixBridge.ts
