+++ src/lib/matrixUtil.js
@@ -1,19 +1,22 @@
-"use strict";
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
+var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    return new (P || (P = Promise))(function (resolve, reject) {
+        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
+        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
+        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
+        step((generator = generator.apply(thisArg, _arguments || [])).next());
+    });
 };
-Object.defineProperty(exports, "__esModule", { value: true });
-const matrix_js_sdk_1 = __importDefault(require("matrix-js-sdk"));
-const debug_1 = __importDefault(require("debug"));
-const debug = debug_1.default("matrixUtil");
-const getSyncMatrixClient = async ({ user = process.env.CYPHERNODE_MATRIX_USER, password = process.env.CYPHERNODE_MATRIX_PASS, baseUrl = process.env.CYPHERNODE_MATRIX_SERVER } = {}) => {
+import matrix from "matrix-js-sdk";
+import _debug from "debug";
+const debug = _debug("matrixUtil");
+const getSyncMatrixClient = ({ user = process.env.CYPHERNODE_MATRIX_USER, password = process.env.CYPHERNODE_MATRIX_PASS, baseUrl = process.env.CYPHERNODE_MATRIX_SERVER } = {}) => __awaiter(this, void 0, void 0, function* () {
     debug("Conneting to", baseUrl, user);
-    const matrixClient = await matrix_js_sdk_1.default.createClient({
+    const matrixClient = yield matrix.createClient({
         baseUrl,
         initialSyncLimit: 10,
         timelineSupport: true
     });
-    await matrixClient.login("m.login.password", {
+    yield matrixClient.login("m.login.password", {
         user,
         password
     });
@@ -52,5 +55,5 @@ const getSyncMatrixClient = async ({ user = process.env.CYPHERNODE_MATRIX_USER, 
             }
         });
     });
-};
-exports.getSyncMatrixClient = getSyncMatrixClient;
+});
+export { getSyncMatrixClient };
