+++ src/lib/matrixUtil.ts
